+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Differences Between DuMu<sup>x</sup> 3.4 and DuMu<sup>x</sup> 3.3
     - Advection type including inertia effect for simulations in the non-creeping flow regime
     - Note that this is still considered a rather _experimental_ feature. Everything within namespace `Dumux::PoreNetwork` might undergo (backward-compatibility breaking) changes _without prior notice_.
 - __Several scripts have been translated to Python__:
+    - `installexternal.sh` to install external modules and libraries now rewritten as python script `bin/installexternal.py`
     - `getusedversions.sh` to extract the used Dumux/Dune versions of a module (new script: `bin/util/getusedversions.py`)
     - `extractmodulepart.sh` no longer creates an install file, instead, you can now generate install scripts for your module using the new script `bin/util/makeinstallscript.py`.
     - Note: the old shells script will be removed after release 3.4.